修改商家继承关系，允许商家账号进行提交订单等相关操作，修复配置读取类自动设置监听地址格式

+++ ManagerProxy/AuthenticationCarrier/AuthenticationCarrier.h
@@ -15,7 +15,7 @@ public:
 	/// @brief Send verification to given code
 	/// @param sendTo: represent where does the code go
 	/// @return the code just sent
-	virtual long sendVerificationCode(std::string sendTo) = 0;
+	virtual unsigned sendVerificationCode(std::string sendTo) = 0;
 };
